refactor:Remove legacy extension and extract runscriptclass
