+++ UIView.ui
@@ -43,40 +43,6 @@
      <property name="bottomMargin">
       <number>0</number>
      </property>
-     <item row="2" column="1">
-      <widget class="QComboBox" name="choice_stop">
-       <property name="sizePolicy">
-        <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
-         <horstretch>0</horstretch>
-         <verstretch>0</verstretch>
-        </sizepolicy>
-       </property>
-      </widget>
-     </item>
-     <item row="4" column="1">
-      <widget class="QComboBox" name="choice_extension">
-       <property name="sizePolicy">
-        <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
-         <horstretch>0</horstretch>
-         <verstretch>0</verstretch>
-        </sizepolicy>
-       </property>
-      </widget>
-     </item>
-     <item row="2" column="0">
-      <widget class="QLabel" name="label_stop">
-       <property name="text">
-        <string>Terminate</string>
-       </property>
-      </widget>
-     </item>
-     <item row="3" column="0">
-      <widget class="QLabel" name="label_language">
-       <property name="text">
-        <string>Language</string>
-       </property>
-      </widget>
-     </item>
      <item row="0" column="1">
       <widget class="QComboBox" name="choice_start">
        <property name="sizePolicy">
@@ -87,6 +53,47 @@
        </property>
        <property name="currentText">
         <string/>
+       </property>
+      </widget>
+     </item>
+     <item row="3" column="1">
+      <widget class="QComboBox" name="choice_language">
+       <property name="sizePolicy">
+        <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
+         <horstretch>0</horstretch>
+         <verstretch>0</verstretch>
+        </sizepolicy>
+       </property>
+      </widget>
+     </item>
+     <item row="2" column="1">
+      <widget class="QComboBox" name="choice_stop">
+       <property name="sizePolicy">
+        <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
+         <horstretch>0</horstretch>
+         <verstretch>0</verstretch>
+        </sizepolicy>
+       </property>
+      </widget>
+     </item>
+     <item row="3" column="0">
+      <widget class="QLabel" name="label_language">
+       <property name="text">
+        <string>Language</string>
+       </property>
+      </widget>
+     </item>
+     <item row="0" column="0">
+      <widget class="QLabel" name="label_start_key">
+       <property name="text">
+        <string>Launch/Pause</string>
+       </property>
+      </widget>
+     </item>
+     <item row="2" column="0">
+      <widget class="QLabel" name="label_stop">
+       <property name="text">
+        <string>Terminate</string>
        </property>
       </widget>
      </item>
@@ -104,30 +111,6 @@
       <widget class="QLabel" name="label_record">
        <property name="text">
         <string>Record/Pause</string>
-       </property>
-      </widget>
-     </item>
-     <item row="3" column="1">
-      <widget class="QComboBox" name="choice_language">
-       <property name="sizePolicy">
-        <sizepolicy hsizetype="Minimum" vsizetype="Minimum">
-         <horstretch>0</horstretch>
-         <verstretch>0</verstretch>
-        </sizepolicy>
-       </property>
-      </widget>
-     </item>
-     <item row="4" column="0">
-      <widget class="QLabel" name="label_extension">
-       <property name="text">
-        <string>Extension</string>
-       </property>
-      </widget>
-     </item>
-     <item row="0" column="0">
-      <widget class="QLabel" name="label_start_key">
-       <property name="text">
-        <string>Launch/Pause</string>
        </property>
       </widget>
      </item>
@@ -196,7 +179,7 @@
       <x>10</x>
       <y>260</y>
       <width>631</width>
-      <height>161</height>
+      <height>216</height>
      </rect>
     </property>
     <layout class="QVBoxLayout" name="verticalLayout">
