+++ requirements-windows.txt
@@ -7,3 +7,4 @@ future==0.18.2
 PySide2==5.15.2.1
 qt-material==2.11
 loguru==0.6.0
+json5==0.9.10
